fix: correction to liver blood flow values
Fb changed from 3.61/60 to 2.27/60 on 07/03/2022. (Notes from Nicola Melillo: 'From Brown the cardiac output of rats is 110.4 mL/min (table 3-1), that is around 6.62L/h. From table 3-4 we see that the sum of hepatic artery and portal vein blood flow is 17.4% of the total cardiac output, that is around 1.152 L/h. Mass of the liver is 9.15g, with a density of 1.08 kg/L, therefore ~8.47mL. 9.18g refers to the whole liver, i.e. intracellular tissue + extracellular space + blood. If we divide 1.152L/h for 8.47mL we obtain ~2.27 mL/h/mL liver.')